# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PMS Hotel Management System API
  description: |
    Complete REST API documentation for the Property Management System.
    
    ## Authentication
    All endpoints (except login) require a Bearer token in the Authorization header.
    
    ## Base URL
    - Development: http://localhost:3001/api
    - Production: https://api.yourhotel.com/api
    
    ## Error Codes
    - 400 - Bad Request
    - 401 - Unauthorized
    - 403 - Forbidden
    - 404 - Not Found
    - 500 - Internal Server Error
  version: 1.0.0
  contact:
    name: API Support
    email: support@grandview.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.grandview.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /payments:
    get:
      tags:
        - Payments
      summary: Get all payments
      security:
        - bearerAuth: []
      parameters:
        - name: invoiceId
          in: query
          schema:
            type: integer
        - name: reservationId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of payments
    post:
      tags:
        - Payments
      summary: Process payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - paymentMethod
                - amount
              properties:
                invoiceId:
                  type: integer
                reservationId:
                  type: integer
                paymentMethod:
                  type: string
                amount:
                  type: number
                currency:
                  type: string
                referenceNumber:
                  type: string
                transactionId:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Payment processed
  /payments/{id}:
    get:
      tags:
        - Payments
      summary: Get payment by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Payment details
  /payments/{id}/refund:
    post:
      tags:
        - Payments
      summary: Process refund
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refundAmount
                - reason
              properties:
                refundAmount:
                  type: number
                reason:
                  type: string
      responses:
        '200':
          description: Refund initiated
  /payments/methods/list:
    get:
      tags:
        - Payments
      summary: Get payment methods
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of payment methods
    post:
      tags:
        - Payments
      summary: Create payment method
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - methodName
              properties:
                methodName:
                  type: string
                methodType:
                  type: string
      responses:
        '201':
          description: Payment method created
  /payments/history:
    get:
      tags:
        - Payments
      summary: Get payment history
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Payment history

  # ============================================================
  # SERVICES
  # ============================================================
...
  /reports/payments:
    get:
      tags:
        - Reports
      summary: Get payment report
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: paymentMethod
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Payment data
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get notifications
      security:
        - bearerAuth: []
      parameters:
        - name: guestId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of notifications
    post:
      tags:
        - Notifications
      summary: Create notification
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - channel
                - body
              properties:
                guestId:
                  type: integer
                reservationId:
                  type: integer
                type:
                  type: string
                channel:
                  type: string
                subject:
                  type: string
                body:
                  type: string
                scheduledDate:
                  type: string
      responses:
        '201':
          description: Notification created
  /notifications/{id}:
    put:
      tags:
        - Notifications
      summary: Update notification
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        '200':
          description: Notification updated
  /notifications/email-templates:
    get:
      tags:
        - Notifications
      summary: Get email templates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of email templates
    post:
      tags:
        - Notifications
      summary: Create email template
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - templateName
                - templateType
                - subject
                - body
              properties:
                templateName:
                  type: string
                templateType:
                  type: string
                subject:
                  type: string
                body:
                  type: string
                variables:
                  type: string
      responses:
        '201':
          description: Template created
  /notifications/email-templates/{id}:
    put:
      tags:
        - Notifications
      summary: Update email template
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                templateName:
                  type: string
                subject:
                  type: string
                body:
                  type: string
                isActive:
                  type: boolean
      responses:
        '200':
          description: Template updated
  /notifications/sms-templates:
    get:
      tags:
        - Notifications
      summary: Get SMS templates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of SMS templates
    post:
      tags:
        - Notifications
      summary: Create SMS template
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - templateName
                - body
              properties:
                templateName:
                  type: string
                templateType:
                  type: string
                body:
                  type: string
                variables:
                  type: string
      responses:
        '201':
          description: Template created
  /notifications/sms-templates/{id}:
    put:
      tags:
        - Notifications
      summary: Update SMS template
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                templateName:
                  type: string
                body:
                  type: string
                isActive:
                  type: boolean
      responses:
        '200':
          description: Template updated